The animated leader token moves to the one page that renders it
K4's first item, and the plain application of the rule. Kind is judged at
the membrane, and this had no membrane: no props at all, both the portrait
list and the edit list baked in, and one importer. There was nothing for a
category to be judged at, so it was never Content in the first place.

A single-caller view component belongs in its caller's file, and the home
route already keeps HomepagePending and HomepageError there. The
stylesheet moves into the page's, with names prefixed because that file
carries the whole page rather than one component.

The story goes with it. A route-local component has no story in this repo,
and the rule requiring one applies to the kit it has now left.

Verified against the running page rather than the diff. Same 320px circle,
same aspect ratio, same drop shadow, same accessible name.

The animation needed working for, because the browser reports
prefers-reduced-motion and the component correctly freezes under it, which
looks identical to a broken timer. Patching the query and remounting shows
the state machine intact: it had cycled off the first leader and was
typing the next name a letter at a time, F, Fa, Far, Faro, Farok.
