+++ src/app/routes/_app/index.module.css
@@ -126,3 +126,60 @@
     min-height: 20rem;
   }
 }
+
+/* The animated leader token, which this page owns. Names are prefixed because this stylesheet holds
+   the whole page rather than one component. */
+.leaderToken {
+  position: relative;
+  width: min(100%, 20rem);
+  aspect-ratio: 1;
+  overflow: hidden;
+  border-radius: 50%;
+  filter: drop-shadow(0 18px 22px rgb(54 29 8 / 28%));
+}
+
+.leaderTokenStable,
+.leaderTokenPrevious,
+.leaderTokenCurrent {
+  position: absolute;
+  inset: 0;
+}
+
+.leaderTokenPrevious {
+  animation: leader-token-fade-out 850ms ease-in-out forwards;
+}
+
+.leaderTokenCurrent {
+  animation: leader-token-fade-in 850ms ease-in-out forwards;
+}
+
+@keyframes leader-token-fade-in {
+  from {
+    opacity: 0;
+    transform: scale(0.985);
+  }
+
+  to {
+    opacity: 1;
+    transform: scale(1);
+  }
+}
+
+@keyframes leader-token-fade-out {
+  from {
+    opacity: 1;
+    transform: scale(1);
+  }
+
+  to {
+    opacity: 0;
+    transform: scale(1.015);
+  }
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .leaderTokenPrevious,
+  .leaderTokenCurrent {
+    animation-duration: 1ms;
+  }
+}
