--- src/app/ui/content/AnimatedLeaderToken.module.css
@@ -1,54 +1,0 @@
-.root {
-  position: relative;
-  width: min(100%, 20rem);
-  aspect-ratio: 1;
-  overflow: hidden;
-  border-radius: 50%;
-  filter: drop-shadow(0 18px 22px rgb(54 29 8 / 28%));
-}
-
-.stable,
-.previous,
-.current {
-  position: absolute;
-  inset: 0;
-}
-
-.previous {
-  animation: fade-out 850ms ease-in-out forwards;
-}
-
-.current {
-  animation: fade-in 850ms ease-in-out forwards;
-}
-
-@keyframes fade-in {
-  from {
-    opacity: 0;
-    transform: scale(0.985);
-  }
-
-  to {
-    opacity: 1;
-    transform: scale(1);
-  }
-}
-
-@keyframes fade-out {
-  from {
-    opacity: 1;
-    transform: scale(1);
-  }
-
-  to {
-    opacity: 0;
-    transform: scale(1.015);
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .previous,
-  .current {
-    animation-duration: 1ms;
-  }
-}
